+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.8)
-project(cpp_pubsub)
+project(ros2_espeaker)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
@@ -24,10 +24,10 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 endif()
 
-add_executable(talker src/publisher_member_function.cpp)
+add_executable(talker src/string_message_publisher.cpp)
 ament_target_dependencies(talker rclcpp std_msgs)
 
-add_executable(listener src/subscriber_member_function.cpp)
+add_executable(listener src/espeak_subscriber_core.cpp)
 ament_target_dependencies(listener rclcpp std_msgs)
 
 install(TARGETS
