--- src/subscriber_member_function.cpp
@@ -1,62 +1,0 @@
-// Copyright 2016 Open Source Robotics Foundation, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include <functional>
-#include <memory>
-
-#include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/string.hpp"
-#include <iostream>
-#include <string.h>
-#include <typeinfo>
-using std::placeholders::_1;
-
-const char* num = "10";
-
-class MinimalSubscriber : public rclcpp::Node
-{
-public:
-  MinimalSubscriber()
-  : Node("minimal_subscriber")
-  {
-    subscription_ = this->create_subscription<std_msgs::msg::String>(
-      "topic", 10, std::bind(&MinimalSubscriber::topic_callback, this, _1));
-  }
-
-private:
-  void topic_callback(const std_msgs::msg::String & msg) const
-  {
-// char[] c = str.toCharArray();
-    if (msg.data == num){
-      std::cout << "yes" << std::endl;
-      std::string st = "espeak ' " + std::string(num) + "'";
-      system(st.c_str());
-    }
-    // int ret = (msg.data.c_str(), "10");
-    std::cout << msg.data.c_str() << std::endl;
-    // RCLCPP_INFO(this->get_logger(), "I heard: '%s'", typeid(msg.data).name());
-      // return 0;
-    // }
-    
-  }
-  rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
-};
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<MinimalSubscriber>());
-  rclcpp::shutdown();
-  return 0;
-}
